requirements.txt
pillow
torch
torchvision
flask
prometheus_client